[LLVM 17] Cast to intptr_t instead of int in dynamic_stack_buffer_overflow_ex3_error_message.cpp

+++ compiler-rt/test/asan/TestCases/Windows/dynamic_stack_buffer_overflow_ex3_error_message.cpp
@@ -24,13 +24,13 @@ int main()
 __try{
   tmp = k;
   aa = (int*)_alloca(SIZE*sizeof(int));
-  if (((int)aa) & 0x3)
+  if (((intptr_t)aa) & 0x3)
     fail = 1;
   for (i=0;i<SIZE;i++){
     aa[i] = x+1+i;
   }
   bb = (int*)_alloca(x*sizeof(int));
-  if (((int)bb) & 0x3)
+  if (((intptr_t)bb) & 0x3)
     fail = 1;
 
   for (i=0;i<x;i++){
@@ -45,7 +45,7 @@ __try{
   }
 
   cc = (int*)_alloca(x);
-  if (((int)cc) & 0x3)
+  if (((intptr_t)cc) & 0x3)
     fail = 1;
 
   cc[0] = 0;
